setup resource caching
