+++ src/utils/mod.rs
@@ -3,5 +3,6 @@ mod file;
 #[cfg(feature = "web")]
 mod metrics;
 mod serde;
-
+mod liquidity_pair;
 pub use serde::*;
+pub use liquidity_pair::*;
